restructure HomeHeader components folder
