+++ frontend/app/(home)/components/HomeHeader/components/QuivrLogo.tsx
